Integrating language switching basics
This commit integrates the existing language switching plumbing with
Banner and other components.
